refactor: Simulation -> SimulationInput 타입 이름 변경
