+++ frontend/src/hooks/simulation/types.ts
@@ -1,0 +1,11 @@
+export type SimulationInput = {
+  id: number;
+  input: string;
+};
+
+export type SimulationResult = {
+  id: number;
+  isDone: boolean;
+  input: string;
+  output: unknown;
+};
